+++ src/main/java/com/codurance/retropolis/requests/NewCardRequestObject.java
@@ -10,7 +10,7 @@ public class NewCardRequestObject {
   private String text;
 
   @NotNull(message = "Column id cannot be empty")
-  private Integer columnId;
+  private Long columnId;
 
   @NotNull(message = "Username cannot be null")
   private String username;
@@ -18,7 +18,7 @@ public class NewCardRequestObject {
   public NewCardRequestObject() {
   }
 
-  public NewCardRequestObject(String text, Integer columnId, String userName) {
+  public NewCardRequestObject(String text, Long columnId, String userName) {
     this.text = text;
     this.columnId = columnId;
     this.username = userName;
@@ -28,7 +28,7 @@ public class NewCardRequestObject {
     return text;
   }
 
-  public int getColumnId() {
+  public Long getColumnId() {
     return columnId;
   }
